update ass2 7

+++ Assignment2/src/academy/pocu/comp2500/assignment2/Cart.java
@@ -21,8 +21,7 @@ public class Cart {
         return this.list
                 .stream()
                 .map(e -> e.getPrice())
-                .reduce(0, (total, e) -> total + e)
-                .intValue();
+                .reduce(0, (total, e) -> total + e);
 
     }
